Remove player Info data from index into it's own components. Needed to move Season Selector there as well in order for data loading to work on the dropdown
